+++ pkg/util/util.go
@@ -1,6 +1,7 @@
 package util
 
 import (
+	"fmt"
 	"sort"
 
 	"github.com/facette/natsort"
@@ -101,4 +102,19 @@ func MustNewLabelsRequirement(key string, op selection.Operator, vals []string) 
 func Now() *metav1.Time {
 	ts := metav1.Now()
 	return &ts
+}
+
+func StatefulSetPodFQDN(namespace, statefulSetName string, ordinal int, serviceName string) string {
+	// The DNS entry for a pod of a stateful set is
+	// $(statefulset name)-(ordinal).$(service name).$(namespace).svc.cluster.local
+	// https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#stable-network-id
+	//
+	// Ending that with a trailing "." to signify an absolute domain name
+	// https://datatracker.ietf.org/doc/html/rfc1034#section-3.1
+	return fmt.Sprintf("%v-%v.%v.%v.svc.cluster.local.",
+		statefulSetName,
+		ordinal,
+		serviceName,
+		namespace,
+	)
 }